addin user edit

+++ app/Http/Controllers/CompanyController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Company;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Http\Request;
+use App\Models\Role;
 use App\User;
 use Image;
 use File;
@@ -109,7 +111,49 @@ class CompanyController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+        request()->validate([
+        'company_name'  => 'required',
+        'company_email' => 'required',
+        'user_id'       => 'required',
+        'status'        => 'required',
+        ]);
+        
+        $company        = Company::find($id);
+        $company->company_name      = $request->company_name;
+        $company->company_email      = $request->company_email;
+        $company->user_id      = $request->user_id;
+        $company->company_telephone      = $request->company_telephone;
+        $company->company_location      = $request->company_location;
+        $company->company_description      = $request->company_description;
+        $company->company_bio      = $request->company_bio;
+        $company->status      = $request->status;
+        $company->save();
+
+
+        if ($request->hasFile('company_logo')) {
+
+            $pathToImage = public_path('files/companies/images/').$company->company_logo;
+            File::delete($pathToImage);
+
+            $project_image = $request->file('company_logo');
+            $filename = time() . '.' . $project_image->getClientOriginalExtension();
+            Image::make($project_image)->resize(340, 340)->save( public_path('/files/companies/images/' . $filename) );
+            
+            $company = Company::where('id',$id)->get()->first();
+            $company->company_logo = $filename;
+            $company->save();
+        }
+
+        $user = User::find($request->user_id);
+        $user->role = 'company-admin';
+        $user->save();
+
+        DB::table('role_user')->where('user_id',$request->user_id)->delete();
+        $user->attachRole(Role::where('name','company-admin')->first());
+
+        $company = Company::where('company_email',$request->company_email)->first();
+
+        return redirect()->route('companies.show',$company->id)->with('success','Company profile created successfully.');
     }
 
     /**
@@ -121,14 +165,18 @@ class CompanyController extends Controller
     public function destroy($id)
     {
         $item = Company::find($id);
-        $item->delete();
 
-        $user = User::find($id);
+        $user = User::find($item->user_id);
         $user->role = 'subscriber';
         $user->save();
 
-        DB::table('role_user')->where('user_id',$id)->delete();
+        DB::table('role_user')->where('user_id',$item->user_id)->delete();
         $user->attachRole(Role::where('name','subscriber')->first());
+
+        $pathToImage = public_path('files/companies/images/').$item->company_logo;
+        File::delete($pathToImage);
+
+        $item->delete();
         
         return redirect()->route('companies.index')->with('danger', 'Company deleted successfully');
     }
